Rename files to match component names
