+++ client/src/components/PageFooter.vue
