+++ client/src/router.js
@@ -1,7 +1,7 @@
 import { createRouter, createWebHistory } from 'vue-router';
-import HomePage from './views/Home.vue';
-import ProfilePage from './views/Profile.vue';
-import AboutPage from './views/About.vue';
+import HomePage from './views/HomePage.vue';
+import ProfilePage from './views/ProfilePage.vue';
+import AboutPage from './views/AboutPage.vue';
 
 export default createRouter({
   mode: 'history',
